update utils component extracting helper functions into single files in a new golder AList, update all extracted testing files, refactored CityList improving the usage of helper components
