+++ components/exercises/chapter-4-oscar/AList/CityList.test.tsx
@@ -12,6 +12,16 @@ describe('ListItem', () => {
     expect(renderedItem).toBeInTheDocument();
   });
 
+  it('should render "No more cities to visit" if the rest of the list is null', () => {
+    const item = {
+      value: 'Item 1',
+      rest: null,
+    };
+    render(<ListItem item={item} />);
+    const renderedItem = screen.getByText('No more cities to visit');
+    expect(renderedItem).toBeInTheDocument();
+  });
+
   it('should render the rest of the list if it exists', () => {
     const item = {
       value: 'Item 1',
@@ -21,19 +31,7 @@ describe('ListItem', () => {
       },
     };
     render(<ListItem item={item} />);
-    const renderedItem1 = screen.getByText('Item 1');
-    const renderedItem2 = screen.getByText('Item 2');
-    expect(renderedItem1).toBeInTheDocument();
-    expect(renderedItem2).toBeInTheDocument();
-  });
-
-  it('should render "No more cities to visit" if the rest of the list is null', () => {
-    const item = {
-      value: 'Item 1',
-      rest: null,
-    };
-    render(<ListItem item={item} />);
-    const renderedItem = screen.getByText('No more cities to visit');
+    const renderedItem = screen.getByText('Item 2');
     expect(renderedItem).toBeInTheDocument();
   });
 });
@@ -65,5 +63,22 @@ describe('CityList component', () => {
     // Use a more specific query to get the list items
     const cityListItems = screen.getAllByRole('listitem');
     expect(cityListItems.length).toBeGreaterThan(0);
+  });
+
+  test('does not add a city if it already exists in the list', () => {
+    render(<CityList />);
+    const addButton = screen.getByRole('button', { name: '+' });
+    fireEvent.click(addButton);
+
+    // Use a more specific query to get the list items
+    const cityListItems = screen.getAllByRole('listitem');
+    const existingCity = cityListItems[cityListItems.length - 1].textContent;
+    fireEvent.click(addButton);
+
+    // Check if the existing city is still the last item in the list
+    const updatedCityListItems = screen.getAllByRole('listitem');
+    const lastCity =
+      updatedCityListItems[updatedCityListItems.length - 1].textContent;
+    expect(lastCity).toBe(existingCity);
   });
 });
